Revert "Renamed package and added md file"
This reverts commit 42c6f97cbde39cc0a5e0bacf34f86a32ded4f4aa.
