--- java-shallow-deep-copy/README.md
@@ -1,7 +1,0 @@
-## Java Shallow Copy Vs Deep Copy
-
-This module contains articles about the differences between shallow copy and deep copy in Java.
-
-### Relevant articles
-
-- [Creating a Deep vs Shallow Copy Of An Object in Java](https://drafts.baeldung.com/deep-shallow-copy-java/)
