+++ java-shallow-deep-copy/src/main/java/com/baeldung/java_shallow_deep_copy/App.java
@@ -1,4 +1,4 @@
-package com.baeldung;
+package com.baeldung.java_shallow_deep_copy;
 
 /**
  * Hello world!
