+++ java-shallow-deep-copy/src/main/java/com/baeldung/java_shallow_deep_copy/data/BankAccount.java
@@ -1,4 +1,4 @@
-package com.baeldung.data;
+package com.baeldung.java_shallow_deep_copy.data;
 
 public class BankAccount implements Cloneable {
