+++ java-shallow-deep-copy/src/main/java/com/baeldung/java_shallow_deep_copy/data/BankAccountShallow.java
@@ -1,4 +1,4 @@
-package com.baeldung.data;
+package com.baeldung.java_shallow_deep_copy.data;
 
 public class BankAccountShallow extends BankAccount {
