+++ java-shallow-deep-copy/src/test/java/com/baeldung/java_shallow_deep_copy/unit/DeepCopyTest.java
@@ -1,11 +1,11 @@
-package com.baeldung.unit;
+package com.baeldung.java_shallow_deep_copy.unit;
 
 import static org.junit.jupiter.api.Assertions.assertNotEquals;
 
 import org.junit.jupiter.api.Test;
 
-import com.baeldung.data.Balance;
-import com.baeldung.data.BankAccountDeep;
+import com.baeldung.java_shallow_deep_copy.data.Balance;
+import com.baeldung.java_shallow_deep_copy.data.BankAccountDeep;
 
 public class DeepCopyTest {
